Refactor acoustic wave module, extract test support, improve stability tests
